Refactor: Image Asset
