--- src/app/page.style.ts
@@ -1,26 +1,0 @@
-import styled from "@emotion/styled";
-
-
-export const container = styled.div`
-    display: flex;
-    width: 100vw;
-    padding: 80px 20px;
-    gap: 10px;
-`
-
-export const MainWrap = styled.div`
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-`
-
-export const MainItem = styled.div`
-    color: #000;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    height: 54px;
-    font-size: 1.5em;
-`
