+++ src/assets/index.ts
@@ -1,1 +1,9 @@
-export const FinhubLogo = "../../../../public/FinHub.svg";
+/** 이미지 관리 파일 **/
+
+export const FinhubLogo = require('../../public/finhub_logo.svg');
+export const SearchIcon = require('../../public/icons/search.svg');
+export const notifyIcon = require('../../public/icons/notify.svg');
+export const PigIcon = require('../../public/icons/pig.svg');
+export const dummyIcon = require('../../public/icons/dummy.svg');
+export const categoryAllIcon = require('../../public/icons/categoryAll.svg');
+export const scrapIcon = require('../../public/icons/scrap.svg');
