+++ tsconfig.json
@@ -20,6 +20,7 @@
     ],
     "baseUrl": ".",
     "paths": {
+      "icons": ["/public/icons/*"],
       "@/*": ["./src/*"],
       "@/app/*": ["./src/app/*"],
       "@/pages/*": ["./src/pages/*"],
